feat: auto-merge
Extract a `github.AutoMerge` component from `NodeProject`. This component will set up mergify to merge approved pull requests which pass CI builds.
