+++ src/github/index.ts
@@ -5,3 +5,4 @@ export * from './workflows';
 export * from './mergify';
 export * from './pr-template';
 export * from './mergify';
+export * from './auto-merge';
